Refactored HTML CSS and removed target=_blank from links
This commit encompasses the refactoring of the HTML CSS to improve the user interface on different hazard example cards by ensuring the image and text are well aligned. The target="_blank" attribute has also been omitted from all hyperlinks in the HTML files, which previously caused each link to open in a new tab, optimizing the user navigation experience.